Add message acknowledgements when applicable.
By default, messages will be double acked (waiting for the ack reply from the server) to reduce the chance for message loss. This can be changed to a single ack by setting the environment variable "JETSTREAM_USE_SINGLE_ACK". Updated versionining to be a minor instead of a patch, which the previous changes should have been.